bin/nxc_manager/pyproject.toml
[tool.poetry]
name = "nxc_manager"
version = "0.1.0"
description = ""
authors = ["xyder <xydr2311@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
plumbum = "^1.7.0"
toml = "^0.10.2"
python-gitlab = "^2.6.0"

[tool.poetry.dev-dependencies]
pylint = "^2.7.0"
ipython = "^7.20.0"
jupyter = "^1.0.0"
ipykernel = "^5.5.0"
python-dotenv = "^0.15.0"
requests = "^2.25.1"
pycryptodome = "^3.10.1"
ipdb = "^0.13.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
